+++ src/sentry/api/urls.py
@@ -546,10 +546,14 @@ from .endpoints.user_index import UserIndexEndpoint
 from .endpoints.user_ips import UserIPsEndpoint
 from .endpoints.user_notification_details import UserNotificationDetailsEndpoint
 from .endpoints.user_notification_fine_tuning import UserNotificationFineTuningEndpoint
-from .endpoints.user_notification_options import UserNotificationOptionsEndpoint
-from .endpoints.user_notification_options_detail import UserNotificationOptionsDetailEndpoint
-from .endpoints.user_notification_providers import UserNotificationProvidersEndpoint
 from .endpoints.user_notification_settings_details import UserNotificationSettingsDetailsEndpoint
+from .endpoints.user_notification_settings_options import UserNotificationSettingsOptionsEndpoint
+from .endpoints.user_notification_settings_options_detail import (
+    UserNotificationSettingsOptionsDetailEndpoint,
+)
+from .endpoints.user_notification_settings_providers import (
+    UserNotificationSettingsProvidersEndpoint,
+)
 from .endpoints.user_organizationintegrations import UserOrganizationIntegrationsEndpoint
 from .endpoints.user_organizations import UserOrganizationsEndpoint
 from .endpoints.user_password import UserPasswordEndpoint
@@ -823,17 +827,17 @@ USER_URLS = [
     ),
     re_path(
         r"^(?P<user_id>[^\/]+)/notification-options/$",
-        UserNotificationOptionsEndpoint.as_view(),
+        UserNotificationSettingsOptionsEndpoint.as_view(),
         name="sentry-api-0-user-notification-options",
     ),
     re_path(
         r"^(?P<user_id>[^\/]+)/notification-options/(?P<notification_option_id>[^\/]+)/$",
-        UserNotificationOptionsDetailEndpoint.as_view(),
+        UserNotificationSettingsOptionsDetailEndpoint.as_view(),
         name="sentry-api-0-user-notification-options-details",
     ),
     re_path(
         r"^(?P<user_id>[^\/]+)/notification-providers/$",
-        UserNotificationProvidersEndpoint.as_view(),
+        UserNotificationSettingsProvidersEndpoint.as_view(),
         name="sentry-api-0-user-notification-providers",
     ),
     re_path(
