+++ tests/sentry/api/endpoints/test_user_notification_settings_options.py
@@ -1,6 +1,6 @@
 from rest_framework import status
 
-from sentry.models.notificationsettingprovider import NotificationSettingProvider
+from sentry.models.notificationsettingoption import NotificationSettingOption
 from sentry.notifications.types import (
     NotificationScopeEnum,
     NotificationSettingEnum,
@@ -8,71 +8,53 @@ from sentry.notifications.types import (
 )
 from sentry.testutils.cases import APITestCase
 from sentry.testutils.silo import control_silo_test
-from sentry.types.integrations import ExternalProviderEnum
 
 
-class UserNotificationProvidersBaseTest(APITestCase):
-    endpoint = "sentry-api-0-user-notification-providers"
+class UserNotificationSettingsOptionsBaseTest(APITestCase):
+    endpoint = "sentry-api-0-user-notification-options"
 
 
 @control_silo_test(stable=True)
-class UserNotificationProvidersGetTest(UserNotificationProvidersBaseTest):
+class UserNotificationSettingsOptionsGetTest(UserNotificationSettingsOptionsBaseTest):
     def setUp(self):
         super().setUp()
         self.login_as(self.user)
 
     def test_simple(self):
         other_user = self.create_user()
-        NotificationSettingProvider.objects.create(
+        NotificationSettingOption.objects.create(
             user_id=self.user.id,
             scope_type=NotificationScopeEnum.ORGANIZATION.value,
             scope_identifier=self.organization.id,
             type=NotificationSettingEnum.ISSUE_ALERTS.value,
-            provider=ExternalProviderEnum.SLACK.value,
             value=NotificationSettingsOptionEnum.ALWAYS.value,
         )
-        NotificationSettingProvider.objects.create(
-            user_id=self.user.id,
-            scope_type=NotificationScopeEnum.ORGANIZATION.value,
-            scope_identifier=self.organization.id,
-            type=NotificationSettingEnum.ISSUE_ALERTS.value,
-            provider=ExternalProviderEnum.EMAIL.value,
-            value=NotificationSettingsOptionEnum.ALWAYS.value,
-        )
-        NotificationSettingProvider.objects.create(
+        NotificationSettingOption.objects.create(
             user_id=self.user.id,
             scope_type=NotificationScopeEnum.ORGANIZATION.value,
             scope_identifier=self.organization.id,
             type=NotificationSettingEnum.WORKFLOW.value,
-            provider=ExternalProviderEnum.EMAIL.value,
             value=NotificationSettingsOptionEnum.ALWAYS.value,
         )
-        NotificationSettingProvider.objects.create(
+        NotificationSettingOption.objects.create(
             user_id=other_user.id,
             scope_type=NotificationScopeEnum.ORGANIZATION.value,
             scope_identifier=self.organization.id,
             type=NotificationSettingEnum.ISSUE_ALERTS.value,
-            provider=ExternalProviderEnum.SLACK.value,
             value=NotificationSettingsOptionEnum.ALWAYS.value,
         )
 
         response = self.get_success_response("me", type="alerts").data
-        assert len(response) == 2
-        slack_item = next(item for item in response if item["provider"] == "slack")
-        email_item = next(item for item in response if item["provider"] == "email")
-
-        assert slack_item["scopeType"] == "organization"
-        assert slack_item["scopeIdentifier"] == str(self.organization.id)
-        assert slack_item["user_id"] == str(self.user.id)
-        assert slack_item["team_id"] is None
-        assert slack_item["value"] == "always"
-        assert slack_item["type"] == "alerts"
-        assert slack_item["provider"] == "slack"
-
-        assert email_item["provider"] == "email"
+        assert len(response) == 1
+        assert response[0]["scopeType"] == "organization"
+        assert response[0]["scopeIdentifier"] == str(self.organization.id)
+        assert response[0]["user_id"] == str(self.user.id)
+        assert response[0]["team_id"] is None
+        assert response[0]["value"] == "always"
+        assert response[0]["type"] == "alerts"
 
         response = self.get_success_response("me").data
-        assert len(response) == 3
+        assert len(response) == 2
 
     def test_invalid_type(self):
         response = self.get_error_response(
@@ -84,7 +66,7 @@ class UserNotificationProvidersGetTest(UserNotificationProvidersBaseTest):
 
 
 @control_silo_test(stable=True)
-class UserNotificationProvidersPutTest(UserNotificationProvidersBaseTest):
+class UserNotificationSettingsOptionsPutTest(UserNotificationSettingsOptionsBaseTest):
     method = "PUT"
 
     def setUp(self):
@@ -100,31 +82,29 @@ class UserNotificationProvidersPutTest(UserNotificationProvidersBaseTest):
             type="alerts",
             status_code=status.HTTP_201_CREATED,
             value="always",
-            provider=["slack"],
         )
-        assert NotificationSettingProvider.objects.filter(
+        row = NotificationSettingOption.objects.filter(
             user_id=self.user.id,
             scope_type=NotificationScopeEnum.ORGANIZATION.value,
             scope_identifier=self.organization.id,
             type=NotificationSettingEnum.ISSUE_ALERTS.value,
             value=NotificationSettingsOptionEnum.ALWAYS.value,
-            provider=ExternalProviderEnum.SLACK.value,
-        ).exists()
-        assert len(response.data) == 3
+        ).first()
+        assert response.data["id"] == str(row.id)
 
     def test_invalid_scope_type(self):
         response = self.get_error_response(
             "me",
             user_id=self.user.id,
-            scope_type="project",
-            scope_identifier=self.project.id,
+            scope_type="invalid",
+            scope_identifier=self.organization.id,
             type="alerts",
             status_code=status.HTTP_400_BAD_REQUEST,
-            provider=["slack"],
+            value="always",
         )
         assert response.data["scopeType"] == ["Invalid scope type"]
 
-    def test_invalid_provider(self):
+    def test_invalid_value(self):
         response = self.get_error_response(
             "me",
             user_id=self.user.id,
@@ -132,6 +112,18 @@ class UserNotificationProvidersPutTest(UserNotificationProvidersBaseTest):
             scope_identifier=self.organization.id,
             type="alerts",
             status_code=status.HTTP_400_BAD_REQUEST,
-            provider=["github"],
+            value="hello",
         )
-        assert response.data["provider"] == ["Invalid provider"]
+        assert response.data["value"] == ["Invalid value"]
+
+    def test_invalid_value_for_option(self):
+        response = self.get_error_response(
+            "me",
+            user_id=self.user.id,
+            scope_type="organization",
+            scope_identifier=self.organization.id,
+            type="alerts",
+            status_code=status.HTTP_400_BAD_REQUEST,
+            value=NotificationSettingsOptionEnum.SUBSCRIBE_ONLY.value,
+        )
+        assert response.data["nonFieldErrors"] == ["Invalid type for value"]
